+++ connector/go.mod
@@ -33,10 +33,10 @@ require (
 	go.opentelemetry.io/otel v1.21.0 // indirect
 	go.opentelemetry.io/otel/metric v1.21.0 // indirect
 	go.opentelemetry.io/otel/trace v1.21.0 // indirect
-	golang.org/x/net v0.19.0 // indirect
+	golang.org/x/net v0.18.0 // indirect
 	golang.org/x/sys v0.15.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20231127180814-3a041ad873d4 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20231002182017-d307bd883b97 // indirect
 	google.golang.org/grpc v1.60.1 // indirect
 	google.golang.org/protobuf v1.31.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
@@ -65,29 +65,3 @@ replace go.opentelemetry.io/collector/receiver => ../receiver
 replace go.opentelemetry.io/collector/exporter => ../exporter
 
 replace go.opentelemetry.io/collector/config/configretry => ../config/configretry
-
-replace go.opentelemetry.io/collector/config/internal => ../config/internal
-
-replace go.opentelemetry.io/collector/config/confighttp => ../config/confighttp
-
-replace go.opentelemetry.io/collector/exporter/otlpexporter => ../exporter/otlpexporter
-
-replace go.opentelemetry.io/collector/config/configtls => ../config/configtls
-
-replace go.opentelemetry.io/collector/extension/auth => ../extension/auth
-
-replace go.opentelemetry.io/collector/config/configcompression => ../config/configcompression
-
-replace go.opentelemetry.io/collector/semconv => ../semconv
-
-replace go.opentelemetry.io/collector/config/configopaque => ../config/configopaque
-
-replace go.opentelemetry.io/collector/config/confignet => ../config/confignet
-
-replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../receiver/otlpreceiver
-
-replace go.opentelemetry.io/collector/config/configgrpc => ../config/configgrpc
-
-replace go.opentelemetry.io/collector/config/configauth => ../config/configauth
-
-replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../exporter/otlphttpexporter
