+++ exporter/otlpexporter/go.mod
@@ -110,12 +110,10 @@ retract (
 
 replace go.opentelemetry.io/collector/config/configtelemetry => ../../config/configtelemetry
 
-replace go.opentelemetry.io/collector/semconv => ../../semconv
-
-replace go.opentelemetry.io/collector/config/configretry => ../../config/configretry
-
 replace go.opentelemetry.io/collector/config/confighttp => ../../config/confighttp
 
 replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../../receiver/otlpreceiver
 
-replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../otlphttpexporter
+replace go.opentelemetry.io/collector/semconv => ../../semconv
+
+replace go.opentelemetry.io/collector/config/configretry => ../../config/configretry
