+++ internal/e2e/consume_contract_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package test
+package e2e
 
 import (
 	"testing"
